requirements.txt
fastapi[standard]==0.115.12
pydantic-settings
requests
rich==14.0.0
jdatetime==5.2.0
instagrapi==2.2.1
openai==1.82.0
